+++ packages/br-validators/src/index.ts
@@ -1,2 +1,3 @@
 export * from './cnpj.js';
 export * from './cpf.js';
+export * from './cep.js';